Changing the data type of uuid columns to char(36) ASCII.
This will reduce the size to 1/3 compared to UTF-8, which also means
indices will 1/3 the size, and you can hold 3x more index (or 3x more
rows of index) in memory.

The .txt file is actually executed by
/cloudfoundry-identity-common/src/main/java/org/cloudfoundry/identity/uaa/db/mysql/V2_0_9__CleanUpIdColumnTypes.java,
which will continue server start up in case the script fails. The script
has a validation phase which ensures that changes to the actual DB
tables do not take place if any of the id columns contain values that
cannot be coerced to CHAR(36) ASCII.